pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dnasecure"
version = "1.0.0"
authors = [
    {name = "SelfPower Team", email = "lordrichado@gmail.com"},
]
description = "DNA sequence encryption and security using self-power decomposition"
readme = "README.md"
requires-python = ">=3.6"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bioinformatics",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "selfpowerdecomposer>=0.1.1",
    "numpy>=1.19.0",
    "biopython>=1.79",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
dnasecure = "dnasecure.cli:main"

[project.urls]
"Homepage" = "https://github.com/joirichi/dnasecure"
"Bug Tracker" = "https://github.com/joirichi/dnasecure/issues"
